Instrument term_hook as event emitter
Replace existing term_hook with an aggregate hook, and add an
eventEmitter hook to that set. Then update everywhere glpk code runs
(that would potentially log) to switch the current thread's callback
and emitter (info) to the correct struct for the thing executing on the
thread.

For review:

The EventEmitter piece ( https://github.com/TakeScoop/cpp-eventemitter )
Allows you to implement Native C++ code that behaves like an
EventEmitter. Users of a native module will be able to write
code like:

```javascript
foo = new NativeExtension();
foo.on("some event", function(value) {
	console.log(value);
})
```

And inside your native code, you will be able to emit events as they
happen, which will trigger the appropriate callbacks in javascript. The
emit occurs in a separate thread from your AsyncWorker, and so will not
appreciably block the work being done.

As a high-level explanation of EventEmitter. I'll first give a
super-high-level view of v8/Nan:

v8 has a bunch of data structures and functions which are not
thread-safe that can only be touched from the event loop on the "main"
thread. If you want anything to happen on the main thread, you schedule
it via `uv_async_send`, passing a `uv_async_t*` to it with the callback
you want executed on the main thread. Only in the context of that
callback can you do anything that touches v8/javascript.

A Native extension does work in an AsyncWorker. An AsyncWorker executes
the `Execute()` method in a separate thread, and so cannot directly do
anything that interfaces with javascript.

So, to allow native code to interface with javascript callbacks as
events happen, I provide an AsyncQueuedProgressWorker base class, which
relies on a shared ringbuffer to pass events into a callback that runs
on the main thread. The callback then does the work of dequeuing it and
shipping the event into javascript and calling the js callback.

To allow C code to invoke the emitter, I provide a specialization of
AsyncQueuedProgressWorker which provides a C emitter function. I also
provide a specialization which uses a reentrant emitter, for use in
multi-threaded C code.

To review this, start by first reviewing
https://github.com/TakeScoop/cpp-eventemitter

A tour of this change:

GLPK implements something called an environment, ( `src/glpk/env/glpenv.h` )
which has within it the ability to store both a term_hook and info to
pass to that hook. It was originally meant to be thread-local
(`src/glpk/env/tls.c`) but that was not implemented at the time
(presumably due to lack of consistency in compiler support).

So the first thing to review there is the macro expansion for
thread_local I have defined there. I stole it from
https://stackoverflow.com/questions/18298280/how-to-declare-a-variable-as-thread-local-portably/18298965#18298965

To enable the environment, the macro `HAVE_ENV` must be defined, so it
has been added to the gyp files. While this doesn't show up as much of a
change in the diff, there are `#ifdef HAVE_ENV` guards littered throughout
the code that light up a whole host of new code-paths. I have gone
through all of them and didn't spot anything worrisome (to find them, I pulled the nearest version to the
version here and diff'd
( https://ftp.gnu.org/gnu/glpk/glpk-4.55.tar.gz ) upstream does NOT have
the `HAVE_ENV` guards, so they show up prominently in the diff). Rename
src/env/env.h to src/env/glpenv.h to avoid the newfile diffs, and
`find src -iname '*.c' -o -iname '*.h' | xargs perl -pi -e 's/"env\.h"/"glpenv.h"/g'`
to trim the noise. This is, by far, the most difficult part of this review.

Next, ensure that every entrypoint has the env setup (
nodeglpk.hpp:NodeGLPK::TermHookManager#ThreadInitDefaultHooks which is
implicitely called from the CTOR of nodeglpk.hpp:NodeGLPK::TermHookGuard
So basically ensure everything uses TermHookGuard.

Also ensure TermHookGuard do not outlive their the info they set. This
is necessary to maintain thread consistency as different things do work
on potentially the same threads (the main thread in particular), and to
ensure there are no dangling pointers left on any thread.

Every Nan::AsyncWorkers becomes a ReentrantCWorker. Ensure that every
ReentrantCWorker's ExecuteWithEmitter has a TermHookGuard in it and
appropriately scoped to have the correct lifetime.

Also, I switched all of the `bool thread` member variables to atomics
since they are read and written in seperate threads and could have
resulted in a race. Make sure that everywhere they are read is now a
'.load()'. `operator=` is overloaded, so assignment is fine as is. (no
need to require `.store()`)

Also, to build this before cpp-eventemitter is pushed to npmjs, comment
out cpp-eventemitter in package.json, clone
https://github.com/TakeScoop/cpp-eventemitter, and in the
cpp-eventemitter repo run `npm link`; then in this repo, with the same
version of npm, run `npm link cpp-eventemitter`... If you know an
easier way to do this, I'm happy to learn it.