Environment allocation fixup
Enabling HAVE_ENV woke up a code-path that enables memory statistics
and a custom allocator that prepends every allocation with a 32 byte
(yes, byte) allocation descriptor used for record-keeping. And also
enabled per-environment memory cleanup, so even if something leaked, it
would get freed when the env was freed.

Unfortunately, by making the environments thread-local, the design of
this does not work if an object is allocated in one thread, and
deallocated in another, which occurs when when Delete
method is invoked after having done work (which had its own allocations)
in another thread.

So with this change, I add some atomic stats for record-keeping, and I
don't bother with thefree-list at all since there is no way of knowing
that dellocation is safe at env teardown.

I add some macros to enable this new behavior: GLOBAL_MEM_STATS (and
HAVE_ATOMIC to enable the __atomic features; any moderately recent
version of gcc/clang should support this; if not, __sync has been around
for ages, and if all else fails just roll the dice with no fencing at
all)