Invoke intopt callback on v8 thread
In this refactor, I had missed why it had originally been a trampoline.
The callback is a Javascript callback! So it is necessary to invoke the
callback on the v8 thread.

Rather than trampoline on Execute() and WorkComplete() I instead keep
things in Execute(), but async_send to the v8 thread to invoke the
cb_func. I condwait on it to finish, and then continue to Execute()
on the worker thread. This ensures memory allocations on the
worker-thread env remain allocated for the duration of the callback and
Execute(), and obviate the need to treat Intopt specially