Instrument term_hook as event emitter (#3)
**Instrument term_hook as event emitter**

- Replaced existing term_hook with an aggregate hook, and add an
eventEmitter hook to that set. Then update everywhere glpk code runs
(that would potentially log) to switch the current thread's callback
and emitter (info) to the correct struct for the thing executing on the
thread.

- Using https://github.com/TakeScoop/cpp-eventemitter to allow native code to interface with javascript callbacks as events happen. I'm using the reentrant specialization of the AsyncQueuedProgressWorker provided from there for GLPK code emit.

- I have enabled the environment code in GLPK ( `src/glpk/env/glpenv.h` )
which has within it the ability to store both a term_hook and info to
pass to that hook. It was originally meant to be thread-local
(`src/glpk/env/tls.c`) but that was not implemented at the time
(presumably due to lack of consistency in compiler support). So I implemented the macro expansion for thread_local there. I stole it from
https://stackoverflow.com/questions/18298280/how-to-declare-a-variable-as-thread-local-portably/18298965#18298965

- I have implemented a decorator of Nan::AsyncWorker's, and added guards for the thread-info for sync calls. 

- Added 'on' methods on all objects.

- I add some atomic stats for record-keeping, and I
don't bother with the to-be-free-list at all since there is no way of knowing
that dellocation is safe at env teardown.

- I add some macros to enable this new behavior: GLOBAL_MEM_STATS (and
HAVE_ATOMIC to enable the __atomic features; any moderately recent
version of gcc/clang should support this; if not, __sync has been around
for ages, and if all else fails just roll the dice with no fencing at
all)

- Add some unit-tests

- Use my RAII container for to call glp_free_env

- Refactor Intopt to keep use uv_async_send to post to the cb_func instead of bouncing back and forth between threads, which, with HAVE_ENV is a pain. The refactor makes it more obvious what is meant to happen, is safer, and avoids hassles with HAVE_ENV